Remove Trashed Methods

+++ server/app/Http/Controllers/CurrencyController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Setting;
 use App\Models\Currency;
-use App\Requests\CurrencyRequest;
 use Illuminate\Http\Request;
+use App\Requests\CurrencyRequest;
 
 class CurrencyController extends Controller
 {
@@ -68,48 +69,20 @@ class CurrencyController extends Controller
     }
 
 
-    public function moveToTrash(Request $req)
+    public function moveToTrash(Request $req, $id)
     {
         CurrencyRequest::validationId($req);
 
-        $currency = Currency::find($req->id);
+        $currency = Currency::find($id);
 
         if (!$currency) return $this->error('The currency was not found', 404);
 
+        $settings = Setting::where('currency_id', $id)->get();
+
+        if ($settings) return $this->error('The currency cannot be delete because it\'s a default in app settings', 422);
+
         $currency->delete();
 
-        return $this->success($req->id, 'The currency has been moved to the trash successfully');
-    }
-
-
-    public function trashed()
-    {
-        $currencies = Currency::onlyTrashed()->get();
-
-        return $this->success($currencies);
-    }
-
-
-    public function restore(Request $req)
-    {
-        CurrencyRequest::validationId($req);
-
-        $isDone = Currency::onlyTrashed()->where('id', $req->id)->restore();
-
-        if (!$isDone) return $this->error('The currency is not in the trash', 404);
-
-        return $this->success($req->id, 'The currency has been restored successfully');
-    }
-
-
-    public function remove(Request $req)
-    {
-        CurrencyRequest::validationId($req);
-
-        $isDone = Currency::onlyTrashed()->where('id', $req->id)->forceDelete();
-
-        if (!$isDone) return $this->error('The currency is not in the trash', 404);
-
-        return $this->success($req->id, 'The currency has been deleted successfully');
+        return $this->success($id, 'The currency has beendeleted successfully');
     }
 }
